refactor: rename store to state/atoms
